pagination added

+++ app/Http/Controllers/AdminController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Works;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class AdminController extends Controller
 {
@@ -18,8 +19,12 @@ class AdminController extends Controller
         $this->middleware('auth');
     }
 
+
     public function index()
     {
+        if (Auth::user()->role != 3) {
+            return redirect('/');
+        }
 
         return view('admin.main', [
             'ilanlar' => Works::count(),
@@ -29,25 +34,37 @@ class AdminController extends Controller
     }
     public function ilanlar()
     {
+        if (Auth::user()->role != 3) {
+            return redirect('/');
+        }
         $ilanlar = Works::orderBy('created_at', 'desc')->get(['id', 'baslik', 'sehir', 'sektor', 'basvuru_sayisi']);
         return view('admin.ilanlar', ["ilanlar" => $ilanlar]);
     }
 
     public function kullanicilar()
     {
-        //silme yapılacak
+        if (Auth::user()->role != 3) {
+            return redirect('/');
+        }
+
         $users = User::where('role', 1)->orderBy('created_at', 'desc')->get(['id', 'name', 'email']);
         return view('admin.users', ['users' => $users]);
     }
     public function firmalar()
     {
-        //silme yapılacak
+        if (Auth::user()->role != 3) {
+            return redirect('/');
+        }
+
         $employers = User::where('role', 2)->orderBy('created_at', 'desc')->get(['id', 'name', 'email']);
         return view('admin.employers', ['employers' => $employers]);
     }
 
     public function kullaniciSil($id)
     {
+        if (Auth::user()->role != 3) {
+            return redirect('/');
+        }
         User::destroy($id);
         return redirect()->back();
     }
